+++ dockerfiles/dashing/jobs/bim360helper.rb
@@ -1,30 +1,36 @@
+#Methods
 module Field
   #Reads the file "Login" and gets the corresponding login and project tickets
   #OUT: login[login_ticket, project_ticket]
-  def get_tickets ()
-    login = [0, 0, 0]
-    tickets = [0,0]
+  def self.get_tickets ()
+    login = {:username => 0 , :password => 0, :project => 0 }
+    tickets = {:login => 0, :project => 0}
     File.open(File.expand_path("../login", __FILE__ ), "r") do |rf|
-        login[0] = rf.readline.chomp
-        login[1] = rf.readline.chomp
-        login[2] = rf.readline.chomp
+        login[:username] = rf.readline.chomp
+        login[:password] = rf.readline.chomp
+        login[:project] = rf.readline.chomp
     end
-    stream = JSON.parse(RestClient.get("http://bim360field.autodesk.com/api/login", :params => {:username => username, :password => password}))
-    tickets[0] = stream["ticket"]
-    stream = JSON.parse(RestClient.get("http://bim360field.autodesk.com/api/projects", :params => {:ticket => login_ticket}))
+
+    stream = JSON.parse(RestClient.get("http://bim360field.autodesk.com/api/login", :params => {:username => login[:username], :password => login[:password]}))
+
+    tickets[:login] = stream["ticket"]
+    stream = JSON.parse(RestClient.get("http://bim360field.autodesk.com/api/projects", :params => {:ticket => tickets[:login] }))
+
     stream.each do |p|
-      if p["name"] == project
-        tickets[1] = p["project_id"]
+      if p["name"] == login[:project]
+        tickets[:project] = p["project_id"]
       end
     end
+
     return tickets
   end
 
   #Performs the REST call to the BIM 360 Field Database to recieve companies.
   #IN: Tickets from get_tickets
   #OUT: Hash of company hashes sorted by company_id,
-  def get_companies (tickets)
-    stream = JSON.parse(RestClient.get("http://bim360field.autodesk.com/api/companies/", :params => {:ticket => tickets[0], :project_id => tickets[1]}))
+  def self.get_companies (tickets)
+    companies = {total: {:name => 0, :open => 0, :complete => 0, :ready => 0, :closed => 0, :total => 0}, companies: Hash.new({name: 0, open: 0, complete: 0, ready: 0, closed: 0, total: 0})}
+    stream = JSON.parse(RestClient.get("http://bim360field.autodesk.com/api/companies/", :params => {:ticket => tickets[:login], :project_id => tickets[:project]}))
     stream.each do |c|
       companies[:companies][c["company_id"]] = {name: c["name"], open: 0, ready: 0, complete: 0, closed: 0, total: 0}
     end
@@ -34,15 +40,15 @@ module Field
   #Performs the REST call to the BIM 360 Field Database to recieve issues.
   #IN: Tickets from get_tickets
   #OUT: Stream of JSON
-  def get_issues (tickets)
-      stream = JSON.parse(RestClient::Request.execute(method: :get, url: "http://bim360field.autodesk.com/api/get_issues/", timeout: nil, headers: {:params => {:ticket => tickets[0], :project_id => tickets[1]}}))
+  def self.get_issues (tickets)
+      return JSON.parse(RestClient::Request.execute(method: :get, url: "http://bim360field.autodesk.com/api/get_issues/", timeout: nil, headers: {:params => {:ticket => tickets[:login], :project_id => tickets[:project]}}))
   end
 
   #Increments the company issue counts base don type of issues
   #IN: Companies Hash, JSON Stream of issues
   #OUT: Hash of company hashes sorted by company_id with counted issues
-  def issues_company_type_sort (companies)
-    stream.each do |i|
+  def self.issues_company_type_sort (companies, issues)
+    issues.each do |i|
         case i["status"]
           when "Open"
             companies[i[:companies]["company_id"]][:open] += 1
@@ -65,7 +71,7 @@ module Field
 
   #Orders companies based on open issues and dislpays the ones with the most in the widgets
   #IN: Companies hash, Array of text names for widgets
-  def send_issue_counts (companies, widgets)
+  def self.send_issue_counts (companies, widgets)
     companies_array = companies[:companies].sort_by { |k, v| v[:open] }.reverse!
     12.times do |i|
       send_event(widgets[1][i], {title: companies_array[i][1][:name], open: companies_array[i][1][:open], ready: companies_array[i][1][:ready], complete: companies_array[i][1][:complete], closed: companies_array[i][1][:closed] })
@@ -75,7 +81,8 @@ module Field
 
   #Calculates the %Complete for companies and displays them in the leaderboard
   #IN: Companies hash, Text name of leaderboard widget
-  def send_leaders (companies, leaderboard_widget)
+  def self.send_leaders (companies, leaderboard_widget)
+    leaders = Hash.new({value: 0})
     companies[:companies].each do |k, v|
       if v[:total] != 0
         value = (v[:closed] * 100) / v[:total]
@@ -84,5 +91,4 @@ module Field
     end
     send_event(leaderboard_widget, { items: leaders.values })
   end
-
 end
